refactor: rename data components
