+++ typescript/financial-dashboard/src/components/atoms/DataList.tsx
@@ -3,7 +3,7 @@ type Props = {
   onContextChange?: (context: any) => void;
 };
 
-export const List: React.FC<Props> = ({ values }) => {
+export const DataList: React.FC<Props> = ({ values }) => {
   return (
     <ul className="list-disc pl-4">
       {values.map((value, index) => (
